refactor: 活体重复收敛（批2：N1 重建双路径合一 + A18 双实现合一 + K1a 编解码迁出）
- PetApp._build_window 抽取首次创建/切角色公共序列（lib 预传入保顺序铁律）
- other_instances_use_agent 双实现合一为 agent_link 单一 helper（并集语义）
- collision.py 帧编解码迁出为 collision_codec.py，物理域归位；
  保留 re-export 过渡（window.py 引用待批3 迁移）

结构优化批2（STRUCT_PLAN_FINAL.md）
