+++ pet/collision.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""多开桌宠碰撞物理核心与协议编解码（纯 Python 实现，无 Qt 依赖）。
+"""多开桌宠碰撞物理核心（纯 Python 实现，无 Qt 依赖）。
 
 包含：
 1. 圆链/椭圆碰撞检测（Broad-phase AABB + Narrow-phase）
@@ -7,20 +7,19 @@
 3. 冲量求解（恢复系数默认 0.82、切向摩擦 mu=0.08、库仑上限、每质量 9000px/s 限制）
 4. 位置分离（逆质量分摊、每次最多 60% 重叠、min 1px / max 12px、0.5px slop、连续 3 tick 强制完整分离）
 5. 稳定重合方向（两 ID 稳定哈希，禁用随机）
-6. 协议帧解析与编码（4 字节大端长度前缀 + UTF-8 JSON，4096 字节上限超限丢弃）
-7. 水位去重（按 epoch 记录每个 pair 最高已应用 tick）
-8. 多体碰撞冲量合并与迭代分离
+6. 多体碰撞冲量合并与迭代分离
+
+协议帧编解码与水位去重已迁至 collision_codec.py（本模块保留 re-export 兼容）。
 """
 
 from __future__ import annotations
 
 import hashlib
-import json
 import math
 from dataclasses import dataclass
-from typing import Any, Dict, List, Optional, Sequence, Set, Tuple
+from typing import Dict, List, Optional, Sequence, Set, Tuple
 
-# ---- 默认物理与协议常量 ----
+# ---- 默认物理常量 ----
 DEFAULT_RESTITUTION: float = 0.82       # 默认恢复系数
 DEFAULT_FRICTION: float = 0.08          # 默认切向摩擦系数
 DEFAULT_MASS_SCALE: float = 1.0         # 默认质量倍率
@@ -28,8 +27,12 @@ DEFAULT_IMPULSE_CAP: float = 9000.0     # 每单位质量等效冲量上限 (px/
 IMPULSE_MIN_APPROACH_SPEED: float = 80.0  # 低于此接近速度只做位置分离 (px/s)
 DEFAULT_BASE_SCALE: float = 0.72        # 基准缩放
 
-FRAME_MAX_LENGTH: int = 4096            # 单帧最大字节数（含/不含前缀，此处限制载荷<=4096）
-HEADER_SIZE: int = 4                    # 4字节无符号大端整数长度头
+# ---- 协议层兼容再导出（编解码已迁至 collision_codec.py） ----
+# window.py（批 1c 已合入，仍经 collision.WatermarkDeduplicator 引用）与存量
+# 测试经本模块访问编解码符号；新代码应直接从 .collision_codec 导入。
+from .collision_codec import (  # noqa: F401
+    DecodeError, FrameStreamDecoder, WatermarkDeduplicator, encode_frame,
+)
 
 # ---- 状态 Flags 位定义 (plan4 §2.1) ----
 FLAG_VISIBLE: int = 1 << 0              # 1: 可见
@@ -707,121 +710,3 @@ def solve_multi_body_collision(
     for pair_k in force_full_pairs:
         new_overlap_history[pair_k] = 0
     return impulse_list, combined_impulses_by_id, new_overlap_history
-
-
-# ---- 协议帧编解码与水位去重 ----
-
-@dataclass
-class DecodeError:
-    """协议解码错误对象（避免抛异常）。"""
-    reason: str
-    raw_data: bytes = b""
-
-
-def encode_frame(obj: Any) -> bytes:
-    """将 Python 对象编码为 4 字节大端长度前缀 + UTF-8 JSON 字节帧。"""
-    payload = json.dumps(obj, ensure_ascii=False, separators=(",", ":")).encode("utf-8")
-    length = len(payload)
-    header = length.to_bytes(HEADER_SIZE, byteorder="big", signed=False)
-    return header + payload
-
-
-class FrameStreamDecoder:
-    """流式帧解析器，支持粘包与半包解析，超过 4096 字节安全丢弃。"""
-
-    def __init__(self, max_frame_len: int = FRAME_MAX_LENGTH) -> None:
-        self._buffer = bytearray()
-        self.max_frame_len = max_frame_len
-
-    def feed(self, chunk: bytes) -> List[Any | DecodeError]:
-        """喂入字节流，返回解析成功的消息对象列表或 DecodeError 列表。"""
-        if not chunk:
-            return []
-        self._buffer.extend(chunk)
-        results: List[Any | DecodeError] = []
-
-        while True:
-            if len(self._buffer) < HEADER_SIZE:
-                break
-
-            # 读取 4 字节大端长度
-            length = int.from_bytes(self._buffer[:HEADER_SIZE], byteorder="big", signed=False)
-
-            # 超限检查
-            if length > self.max_frame_len or length < 0:
-                dropped = bytes(self._buffer[:HEADER_SIZE])
-                del self._buffer[:HEADER_SIZE]
-                results.append(DecodeError(reason=f"Frame length {length} exceeds limit {self.max_frame_len}", raw_data=dropped))
-                # The payload length is untrusted, so discard only this header
-                # and search the remaining stream for the next plausible header.
-                sync_at = None
-                for offset in range(len(self._buffer) - HEADER_SIZE + 1):
-                    candidate = int.from_bytes(self._buffer[offset:offset + HEADER_SIZE], "big")
-                    if 0 < candidate <= self.max_frame_len:
-                        sync_at = offset
-                        break
-                if sync_at is None:
-                    self._buffer[:] = self._buffer[-(HEADER_SIZE - 1):]
-                    break
-                del self._buffer[:sync_at]
-                continue
-
-            # 空帧处理 (length == 0)
-            if length == 0:
-                # 移除这 4 字节
-                del self._buffer[:HEADER_SIZE]
-                results.append(DecodeError(reason="Empty frame (length 0)", raw_data=b""))
-                continue
-
-            # 检查是否接收完整帧载荷
-            if len(self._buffer) < HEADER_SIZE + length:
-                # 半包，等待更多数据
-                break
-
-            # 提取完整载荷
-            payload_bytes = bytes(self._buffer[HEADER_SIZE:HEADER_SIZE + length])
-            del self._buffer[:HEADER_SIZE + length]
-
-            try:
-                text = payload_bytes.decode("utf-8")
-                obj = json.loads(text)
-                results.append(obj)
-            except UnicodeDecodeError as e:
-                results.append(DecodeError(reason=f"UTF-8 decode error: {e}", raw_data=payload_bytes))
-            except json.JSONDecodeError as e:
-                results.append(DecodeError(reason=f"JSON decode error: {e}", raw_data=payload_bytes))
-
-        return results
-
-
-class WatermarkDeduplicator:
-    """基于 epoch / pair / tick 的水位去重器 (plan4 §2.1 & §3.2)。
-    
-    客户端每个 epoch 内以 pair 为键记录最高已应用 tick 的水位，不重复应用低于或等于水位的事件。
-    当 epoch 变更时，整体重置水位表。
-    """
-
-    def __init__(self) -> None:
-        self.current_epoch: str = ""
-        self.watermarks: Dict[str, int] = {}
-
-    def should_apply(self, epoch: str, pair: str, tick: int) -> bool:
-        """检查该 impulse 是否应当被应用。
-        
-        如果通过，更新水位并返回 True；若已重复或已过期则返回 False。
-        """
-        if not epoch or not pair:
-            return False
-
-        # epoch 切换：整体替换
-        if epoch != self.current_epoch:
-            self.current_epoch = epoch
-            self.watermarks = {pair: tick}
-            return True
-
-        last_tick = self.watermarks.get(pair, -1)
-        if tick > last_tick:
-            self.watermarks[pair] = tick
-            return True
-
-        return False
