+++ pet/uninstall_cleanup.py
@@ -4,34 +4,14 @@
 供 `--uninstall-cleanup` 参数使用——安装包（Inno Setup）卸载时调用。
 该路径不启动 QApplication/事件循环，只做必要的清理，便于「无 Qt 依赖路径」
 （仍可 import Qt 模块，但不建窗口）下执行。
+
+多实例占用判断由 agent_link 模块统一提供（语义并集：当前变体目录 + <base>
+下全部变体任一认为在用即保留），本模块只负责消费。
 """
 
 from __future__ import annotations
 
-import json
-
-
-def other_instances_use_agent(config, agent_key: str) -> bool:
-    """<base> 下其他变体 / 多开实例是否仍开启该 Agent 联动。
-
-    DSH 桥接插件目录与变体无关（<base>/dsh-pet-bridge），卸载前必须确认
-    没有任何变体或实例还在用它，否则保留插件。
-    """
-    base = config.dir.parent
-    try:
-        candidates = list(base.glob("dsh-pet-standalone*/config*.json"))
-    except OSError:
-        return False
-    for f in candidates:
-        if config.path and f == config.path:
-            continue
-        try:
-            data = json.loads(f.read_text(encoding="utf-8"))
-        except Exception:
-            continue
-        if isinstance(data, dict) and bool((data.get("agent_link") or {}).get(agent_key, False)):
-            return True
-    return False
+from .agent_link import ClaudeCodeMonitor, DshMonitor, other_instances_use_agent
 
 
 def run_uninstall_cleanup(config=None) -> dict:
@@ -50,16 +30,14 @@ def run_uninstall_cleanup(config=None) -> dict:
 
     results = {"autostart": bool(autostart.disable())}
 
-    from .agent_link import ClaudeCodeMonitor, DshMonitor
-
-    if other_instances_use_agent(config, "claude"):
-        results["claude_hooks"] = "skipped"
-    else:
-        results["claude_hooks"] = bool(ClaudeCodeMonitor.uninstall_hooks())
-
-    if other_instances_use_agent(config, "dsh"):
-        results["dsh_bridge"] = "skipped"
-    else:
-        results["dsh_bridge"] = bool(DshMonitor.uninstall_bridge())
+    # 其他实例仍在使用对应联动则保留（hooks/桥接插件是全局状态）
+    for agent_key, result_key, uninstaller in (
+        ("claude", "claude_hooks", ClaudeCodeMonitor.uninstall_hooks),
+        ("dsh", "dsh_bridge", DshMonitor.uninstall_bridge),
+    ):
+        if other_instances_use_agent(config, agent_key):
+            results[result_key] = "skipped"
+        else:
+            results[result_key] = bool(uninstaller())
 
     return results
